# web/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mosic 2.0 API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/ruoli_cipe": {
            "get": {
                "tags": [
                    "Ruoli"
                ],
                "summary": "Lista Ruoli interni all'Ufficio",
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "response": 200,
                                "total_results": 8,
                                "limit": "99999",
                                "offset": 0,
                                "data": [
                                    {
                                        "id": 1,
                                        "codice": 0,
                                        "denominazione": ""
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ruoli"
                ],
                "summary": "Creazione ruolo",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            }
        },
        "/api/ruoli_cipe/{id}": {
            "get": {
                "tags": [
                    "Ruoli"
                ],
                "summary": "Singolo Ruolo",
                "operationId": "idRuoli",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del ruolo",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "response": 200,
                                "total_results": 8,
                                "limit": "99999",
                                "offset": 0,
                                "data": [
                                    {
                                        "id": 1,
                                        "codice": 0,
                                        "denominazione": ""
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ruoli"
                ],
                "summary": "Salvataggio ruolo",
                "operationId": "idRuoli",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id del ruolo",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ruoli",
                        "in": "body",
                        "description": "Richiesta",
                        "required": true,
                        "schema": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo",
                        "examples": {
                            "application/json": {
                                "id": 1,
                                "codice": 0,
                                "denominazione": "Documenti di seduta",
                                "descrizione": "Telex, Appunto generale, passi, etc...",
                                "id_uffici": 2
                            }
                        }
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ruoli"
                ],
                "summary": "Eliminazione ruolo",
                "operationId": "idRuoli",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del ruolo",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operazione avvenuta con successo"
                    },
                    "401": {
                        "description": "Autorizzazione negata"
                    },
                    "409": {
                        "description": "Il ruolo \u00e8 associato ad un utente, impossibile eliminarlo."
                    }
                }
            }
        },
...
    },
...
}